fixed to work with binary ninja 3.0 API
